refactor(chat): split CRUD from orchestration in chat/service
